refactor(SharedUtils): extract JSONCoding, PathResolver, Formatting, FTSQuery, SchemaVersion + URL extensions from Shared (#298)
Closes #298. Plan task 1.4.

Splits the small Foundation-layer utilities out of Shared into a new
SharedUtils target. Callers that import Shared keep seeing every type
unchanged, via @_exported import SharedUtils in Shared/Exports.swift.

Files moved:

- Shared/JSONCoding.swift -> SharedUtils/
- Shared/PathResolver.swift -> SharedUtils/ (also picks up
  import SharedConstants for Shared.Constants.X references)
- Shared/Formatting.swift -> SharedUtils/ (extension Shared, picks up
  import SharedConstants for the namespace)
- Shared/FTSQuery.swift -> SharedUtils/ (same pattern)
- Shared/SchemaVersion.swift -> SharedUtils/ (same pattern)

Carried during 1.4 (not in the original plan):

- The URL extensions (expandingTildeInPath, knownGood) move from
  Shared/CupertinoShared.swift to SharedUtils/URLExtensions.swift.
  PathResolver reaches into URL.expandingTildeInPath at runtime;
  keeping the extension in Shared would require SharedUtils -> Shared,
  which cycles with the chosen Shared -> SharedUtils direction. The
  pointer comment in CupertinoShared.swift documents the move.

Files added:

- SharedUtils/URLExtensions.swift (the moved URL extension)
- One line in Shared/Exports.swift: @_exported import SharedUtils

Package.swift:

- New SharedUtils target, deps [SharedConstants].
- New SharedUtils product.
- Shared target deps go from [SharedConstants] to
  [SharedConstants, SharedUtils].

Plan 1.4 amended to note the URL-extension carry.

Verification:

- xcrun swift build (full package, clean): green, 32.89s
- xcrun swift package clean && xcrun swift test: 1293 tests, 144
  suites, 0 failures
- xcrun swift run cupertino --help: green
- swiftformat: 0 / 9 files reformatted
- swiftlint on SharedUtils/: 0 violations, 0 serious

Public API: unchanged. URL.expandingTildeInPath, URL.knownGood, the
Shared.Formatting / Shared.FTSQuery / Shared.SchemaVersion extension
members, and the module-level JSONCoding / PathResolver enums all
resolve through the Shared re-export with no source change.
